# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HMS API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/patients/prescriptions": {
      "get": {
        "summary": "Get prescription history (Self)",
        "tags": [
          "Patients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Prescription history entries"
          }
        }
      }
    },
    "/api/v1/patients/prescriptions/{id}": {
      "get": {
        "summary": "Get a specific patient's prescription history (Doctor/Admin only)",
        "tags": [
          "Patients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of prescriptions"
          }
        }
      },
      "post": {
        "summary": "Create/Add a prescription for a patient (Doctor/Admin only)",
        "tags": [
          "Patients"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionHistory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prescription created successfully"
          },
          "403": {
            "description": "Forbidden - Doctor/Admin only"
          }
        }
      }
    },
...
    "/api/v1/prescriptions": {
      "post": {
        "summary": "Create a digital prescription (Doctor only)",
        "tags": [
          "Prescription"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "patientId",
                  "medicines"
                ],
                "properties": {
                  "patientId": {
                    "type": "string"
                  },
                  "consultationId": {
                    "type": "string"
                  },
                  "medicines": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MedicineInput"
                    }
                  },
                  "generalInstructions": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prescription created and PDF generated"
          },
          "400": {
            "description": "Validation error"
          }
        }
      },
      "get": {
        "summary": "View all prescriptions for the hospital (Pharmacist/Admin)",
        "tags": [
          "Prescription"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of prescriptions"
          }
        }
      }
    },
    "/api/v1/prescriptions/{id}": {
      "put": {
        "summary": "Update an existing prescription (Doctor only)",
        "tags": [
          "Prescription"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "medicines": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MedicineInput"
                    }
                  },
                  "generalInstructions": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prescription updated and PDF regenerated"
          }
        }
      },
      "get": {
        "summary": "View a specific prescription document by ID",
        "tags": [
          "Prescription"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prescription details including PDF link"
          }
        }
      }
    },
    "/api/v1/prescriptions/patient/me": {
      "get": {
        "summary": "View logged-in patient's own prescriptions (Patient only)",
        "tags": [
          "Prescription"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of prescriptions"
          }
        }
      }
    },
    "/api/v1/prescriptions/patient/{patientId}": {
      "get": {
        "summary": "View a patient's prescriptions (Doctor/Admin only)",
        "tags": [
          "Prescription"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "patientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of prescriptions for the patient"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Medicine": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "dosage": {
            "type": "string"
          },
          "frequency": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          }
        }
      },
      "PrescriptionHistory": {
        "type": "object",
        "properties": {
          "doctorName": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "medicines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Medicine"
            }
          },
          "instructions": {
            "type": "string"
          }
        }
      },
...
      "MedicineInput": {
        "type": "object",
        "required": [
          "name",
          "dosage",
          "frequency",
          "duration"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "dosage": {
            "type": "string"
          },
          "frequency": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "instructions": {
            "type": "string"
          }
        }
      }
    }
  },
...
}